+++ example/ios/RunnerTests/RunnerTests.swift
@@ -1,27 +1,16 @@
-import Flutter
-import UIKit
 import XCTest
-
-
 @testable import xlog
 
-// This demonstrates a simple unit test of the Swift portion of this plugin's implementation.
-//
-// See https://developer.apple.com/documentation/xctest for more information about using XCTest.
-
 class RunnerTests: XCTestCase {
+  func testConfigurationDefaults() {
+    let configuration = XLogConfiguration(logDirectory: "/tmp/log", namePrefix: "test")
 
-  func testGetPlatformVersion() {
-    let plugin = XlogPlugin()
-
-    let call = FlutterMethodCall(methodName: "getPlatformVersion", arguments: [])
-
-    let resultExpectation = expectation(description: "result block must be called.")
-    plugin.handle(call) { result in
-      XCTAssertEqual(result as! String, "iOS " + UIDevice.current.systemVersion)
-      resultExpectation.fulfill()
-    }
-    waitForExpectations(timeout: 1)
+    XCTAssertEqual(configuration.logDirectory, "/tmp/log")
+    XCTAssertEqual(configuration.namePrefix, "test")
+    XCTAssertEqual(configuration.level, .debug)
+    XCTAssertEqual(configuration.mode, .async)
+    XCTAssertEqual(configuration.compressMode, .zlib)
+    XCTAssertEqual(configuration.compressLevel, 6)
+    XCTAssertTrue(configuration.consoleLogEnabled)
   }
-
 }